refactor: move Header and Footer to general components
